refactor(payload-documents): consolidate worker endpoints + tighten typing
Extracts the duplication that crept in across the three internal worker
endpoints introduced in 0.4.0:

* `requireInternalSecret` lives in `endpoints/shared.ts` instead of being
  redeclared verbatim in parse-context, parse-result, and parse-file.
* `fetchInternalDocument` consolidates the `findByID` + `overrideAccess: true`
  + try/catch pattern that parse-context and parse-file both did inline; both
  fetchDocument (user access) and fetchInternalDocument (override) now
  delegate to one private `loadDocument` helper.
* `endpoints/inline-helpers.ts` (new) holds `getLlamaParseClient` and
  `fetchUploadedFile` — they're only used by the inline parse + parse-status
  paths and didn't belong in `shared.ts`.
* `loopback URL` rewrite logic in `fetchUploadedFile` extracted to a tiny
  `rewriteToLoopback` helper for readability.
* `DocumentRecord` moves to `plugin/types.ts` so `ResolveFileBinary.doc` can
  use it (hosts now get field autocompletion in their callbacks). The
  internal `WorkerEndpointConfig` alias is dropped in favour of the public
  `DocumentsWorkerConfig`.
* `parse-endpoint.ts` no longer falls back to the now-invalid `'default'`
  LlamaParse mode; passes `undefined` (API picks its own default).

No public-API or behavioural changes. The endpoints + types are tighter,
and adding a fourth internal endpoint won't require copying the auth helper
again.
